api/requirements.txt
asgiref==3.8.1
certifi==2024.12.14
charset-normalizer==3.4.0
coreapi==2.3.3
coreschema==0.0.4
Django==4.2
django-cors-headers==4.6.0
django-redis==5.2.0
djangorestframework==3.14.0
drf-yasg==1.21.4
geographiclib==2.0
geopy==2.3.0
gunicorn==20.1.0
idna==3.10
inflection==0.5.1
iniconfig==2.0.0
itypes==1.2.0
Jinja2==3.1.4
MarkupSafe==3.0.2
packaging==24.2
pluggy==1.5.0
psycopg2==2.9.3
pytest==8.3.4
pytest-django==4.9.0
pytz==2024.2
redis==5.2.1
requests==2.31.0
ruamel.yaml==0.18.6
ruamel.yaml.clib==0.2.12
sqlparse==0.5.3
uritemplate==4.1.1
urllib3==2.2.3
black==23.7.0
flake8==6.0.0
isort==5.12.0
